requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
ChatterBot==1.2.0
chatterbot-corpus==1.2.2
click==8.1.8
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
daphne==4.1.2
defusedxml==0.7.1
diff-match-patch==20241021
distro==1.9.0
Django==4.2.19
django-allauth==65.3.1
django-ckeditor==6.7.2
django-ckeditor-5==0.2.18
django-colorfield==0.14.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
django-import-export==4.3.7
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-maintenance-mode==0.22.0
django-modeladmin-reorder==0.3.1
django-mptt==0.17.0
django-nested-admin==4.1.4
django-redis==5.4.0
djangorestframework==3.15.2
drf-yasg==1.21.9
et_xmlfile==2.0.0
fastapi==0.115.8
filelock==3.18.0
Flask==3.1.0
flask-mysql-connector==1.1.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.2
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.9.0
joblib==1.5.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
mathparse==0.1.2
matplotlib==3.10.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mysql-connector-python==9.1.0
mysqlclient==2.2.7
narwhals==1.37.1
networkx==3.4.2
numpy==2.2.1
odfpy==1.4.1
openai==0.28.0
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
ping3==4.0.8
plotly==6.0.1
prometheus_client==0.21.1
propcache==0.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyactiveresource==2.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyOpenSSL==25.0.0
pyparsing==3.2.1
pysnmp==7.1.20
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-fsutil==0.15.0
python-monkey-business==1.1.0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.2
razorpay==1.4.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==4.1.0
service-identity==24.2.0
setuptools==80.7.1
ShopifyAPI==12.7.0
six==1.17.0
sniffio==1.3.1
speedtest-cli==2.1.3
SQLAlchemy==2.0.38
sqlparse==0.5.3
starlette==0.45.3
sympy==1.14.0
tablib==3.8.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
twilio==9.5.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
wheel==0.45.1
xlrd==2.0.1
xlwt==1.3.0
yarl==1.18.3
zope.interface==7.2
